pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "article-rss-proxy"
version = "0.1.0"
description = "arXiv の論文を取得し、Gemini でフィルタリング・翻訳した上で RSS フィードとして配信するツール"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "オマージュ海老", email = "hommage.ebi@gmail.com" },
]

dependencies = [
    "feedgen==1.0.0",
    "requests>=2.31",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "python-dateutil>=2.9",
    "google-genai>=0.4.0", # Gemini SDK
    "python-dotenv>=1.0",
    "joblib>=1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
